Refactored review fetching into a helper function. Added enforcement of am/pm time presentation over military time in main agent prompt.
