+++ server/src/helpers/assistants/scheduleBuilder/tools.ts
@@ -2,12 +2,7 @@
 import { tool } from "@langchain/core/tools";
 import { ToolMessage } from "@langchain/core/messages";
 import { z } from "zod";
-import {
-  CourseTerm,
-  Section,
-  SelectedSection,
-  ProfessorRatingDocument,
-} from "@polylink/shared/types";
+import { CourseTerm, Section, SelectedSection } from "@polylink/shared/types";
 import { StateAnnotation } from "./state";
 
 // “Injected” helpers -------------------------------------------------
@@ -19,13 +14,13 @@ import {
   removeFromSchedule,
   findSectionsByFilter,
   getSectionsWithPairs,
+  buildSectionSummaries,
 } from "./helpers";
 import { transformSectionToSelectedSection } from "../../../db/models/schedule/transformSection";
 import {
   sectionQueryAssistant,
   SectionQueryResponse,
 } from "../SectionQuery/sectionQueryAssistant";
-import { getProfessorRatings } from "../../../db/models/professorRatings/professorRatingServices";
 import { Command, getCurrentTaskInput } from "@langchain/langgraph";
 import { getScheduleById } from "../../../db/models/schedule/scheduleServices";
 import { getSelectedSectionsByUserId } from "../../../db/models/selectedSection/selectedSectionServices";
@@ -214,78 +209,8 @@ export const fetchSections = tool(
     );
     sectionsToReturn = sectionsWithPairs as SelectedSection[];
 
-    // Get professor ratings for retrieved sections
-    let sectionSummaries: string[] = [];
-    try {
-      const professorIds = Array.from(
-        new Set(
-          (sectionsToReturn as SelectedSection[]).flatMap((sec) =>
-            (sec.professors ?? [])
-              .map((prof) => prof.id)
-              .filter((id): id is string => !!id)
-          )
-        )
-      );
-      const courseIds = Array.from(
-        new Set(
-          (sectionsToReturn as SelectedSection[]).map((sec) => sec.courseId)
-        )
-      );
-
-      const professorRatings = professorIds.length
-        ? await getProfessorRatings(professorIds, courseIds, {
-            firstName: 1,
-            lastName: 1,
-            overallRating: 1,
-            tags: 1,
-          } as unknown as Partial<ProfessorRatingDocument>)
-        : [];
-
-      const ratingMap = new Map<string, Partial<ProfessorRatingDocument>>();
-      professorRatings.forEach((doc) => {
-        if (doc.id) ratingMap.set(doc.id as string, doc);
-      });
-
-      const formatMeetingTimes = (sec: SelectedSection): string => {
-        return sec.meetings
-          .map((m) => {
-            const days = Array.isArray(m.days) ? m.days.join("") : m.days || "";
-            return `${days} ${m.start_time}-${m.end_time}`;
-          })
-          .join("; ");
-      };
-
-      // Create section summaries for the LLM to process
-      sectionSummaries = (sectionsToReturn as SelectedSection[]).map((sec) => {
-        const instStrings = (sec.professors ?? [])
-          .map((prof) => {
-            const rating = prof.id ? ratingMap.get(prof.id) : undefined;
-            const score =
-              rating?.overallRating !== undefined
-                ? rating.overallRating?.toFixed(1)
-                : "N/A";
-            const snippet = // Top 3 tags associated with the professor
-              rating?.tags && Object.keys(rating.tags).length
-                ? ` [${Object.entries(rating.tags)
-                    .sort((a, b) => (b[1] as number) - (a[1] as number))
-                    .slice(0, 3)
-                    .map(([tag]) => tag)
-                    .join(", ")} ]`
-                : "";
-            return `${prof.name} (${score}/4.0 rating) Popular tags:${snippet}`;
-          })
-          .join("; ");
-
-        return `${sec.courseId} – ${sec.classNumber}\n${formatMeetingTimes(sec)}\n${instStrings}`;
-      });
-    } catch (err) {
-      console.error("Error summarizing sections with ratings", err);
-      sectionSummaries = (sectionsToReturn as SelectedSection[]).map(
-        (sec) => `${sec.courseId} – ${sec.classNumber}`
-      );
-    }
-
-    console.log(sectionSummaries);
+    // Build section summaries to return
+    const sectionSummaries = await buildSectionSummaries(sectionsToReturn);
 
     /* ------------------ final payload ------------------ */
     return new Command({
